+++ src/RaylibCircleExt.cpp
@@ -15,7 +15,7 @@ void raylibCircleExt::update_circle_center(Object::ObjPos new_center) {
     _circle_center = new_center;
 }
 
-void raylibCircleExt::set_velocity(int new_velocity) {
+void raylibCircleExt::set_velocity(double new_velocity) {
     _velocity = new_velocity;
 }
