+++ src/RaylibCircleExt.h
@@ -11,12 +11,12 @@ public:
     ~raylibCircleExt();
     void change_direction_of_travel() { _direction_of_travel *= -1; };
     void update_circle_center(Object::ObjPos new_center);
-    void set_velocity(int new_velocity);
+    void set_velocity(double new_velocity);
     void set_direction_of_travel(int direction);
     void set_mass(int mass);
 
     int direction() const { return _direction_of_travel; }
-    int velocity() const { return _velocity; }
+    double velocity() const { return _velocity; }
     int radius() const { return _radius; }
     int mass() const { return _mass; }
     Object::ObjPos center() const { return _circle_center; }
@@ -24,7 +24,7 @@ public:
 private:
     // Left is -1, right is 1
     int _direction_of_travel;
-    int _velocity;
+    double _velocity;
     int _radius;
     int _mass;
     Object::ObjPos _circle_center;
